Tweaked the existing speedup PR to simplify the code
Also set it to calculate the sparse matrix of mutation features only once,
rather than in the loop of batches, so it should be a bit quicker. Simplified
the logic around caching significantly, so it only sets the neighbour list and
a list of new/updated feature profiles. This reduces complexity and duplicated
code. Still have one failing test though, so it needs a bit more work.